Fixes memory exhaustion when downloading big files
Makes parsing the body of a response (as JSON) a lazy operation, rather
than eager. Parsing the body causes the entire response to be loaded
into memory at once, which can lead to memory exhaustion when attempted
on very large responses (e.g., the binary contents of a very large
file).

Fixes #18